% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\alias{notify_google}
\alias{notify_slack}
\alias{notify_pushbullet}
\alias{notify_zulip}
\title{Send a notification}
\usage{
notify_google(msg, config_file = "~/.rgooglespaces", ...)

notify_pushbullet(msg, ...)

notify_slack(msg, config_file = "~/.slackr", ...)

notify_zulip(msg, config_file = "~/.rzulip", ...)
}
\arguments{
\item{msg}{the message to send}

\item{config_file}{character specifying the path to a credentials configuration file}

\item{...}{additional arguments to internal functions}
}
\value{
the response (invisibly)
}
\description{
Send a notification via one of the following message/notification apps:
}
\details{
\itemize{
\item Google Workspace Spaces
\item Pushbullet
\item Slack
\item Zulip
}

See below for more details.
}
\section{Google Workspace Spaces}{


\code{notify_google()} is a wrapper around \code{httr::POST()}

Requires a Google Workspace account and 'Space' be setup and configured with a webhook url:
\enumerate{
\item From Gmail, go to 'Spaces' on the left hand navigation bar;
\item Create a new space or select an existing one;
\item From the space's dropdown menu, select 'Apps & integrations';
\item Click on 'Manage webhooks' band create an incoming webhook;
\item Copy the webhook url;
\item Create a text file \verb{~/.rgooglespaces}, and add the following using your copied webhook url:

\if{html}{\out{<div class="sourceCode">}}\preformatted{webhook_url: https://chat.googleapis.com/v1/spaces/your_key_and_token_etc
}\if{html}{\out{</div>}}
}



\code{notify_google()} is a wrapper around \code{httr::POST()}

Requires Zulip be setup and configured with a webhook url:
\enumerate{
\item Make note of your Zulip domain (e.g., \code{yourZulipDomain.zulipchat.com})
\item Create a Zulip bot of type 'Incoming Webhook' following \url{https://zulip.com/help/add-a-bot-or-integration};
\item Note the bot's API key following \url{https://zulip.com/api/api-keys};
\item Create a text file \verb{~/.rgooglespaces}, and add the following using your bot's API key,
your Zulip domain, and optionally a default stream and topic to post notifications to:

\if{html}{\out{<div class="sourceCode">}}\preformatted{zulip_apikey:
zulip_domain: yourZulipDomain.zulipchat.com
zulip_stream: optional stream name
zulip_topic: optional topic name
}\if{html}{\out{</div>}}
}

If a stream is not specified, private messages will be sent to the creator of the bot.
}

\section{Pushbullet}{


\code{notify_pushbullet()} is a wrapper around \code{RPushbullet::pbPost()}

Use \code{RPushbullet::pbSetup()} to configure your machine to use Pushbullet.
}

\section{Slack}{


\code{notify_slack()} is a wrapper around \code{slackr::slackr_msg()}
}

